Fix Python 3.4-specific NumPy warnings
With NumPy >= 1.8 and Python 3.4, a bunch of extra warnings are
raised when we set the warnings filter to be 'always'. Specifically,
we 1) set size_in and size_out to be np.int64s instead of ints, and
2) try to index into the dimensions of EnsembleArray output nodes with
floats instead of ints. This fixes both of those problems, which
(thanks to these warnings now showing up) don't become errors later on!

+++ nengo/networks/ensemblearray.py
@@ -124,7 +124,7 @@ class EnsembleArray(nengo.Network):
         output = nengo.Node(output=None, size_in=sizes.sum(), label=name)
         setattr(self, name, output)
 
-        indices = np.zeros(len(sizes) + 1)
+        indices = np.zeros(len(sizes) + 1, dtype=int)
         indices[1:] = np.cumsum(sizes)
         for i, e in enumerate(self.ea_ensembles):
             nengo.Connection(
